Add `helper` method to tests
