+++ spec/minitest-rails/spec_type_spec.rb
@@ -6,58 +6,58 @@ describe "minitest-rails spec_type Lookup Integration" do
   context "ProductDecorator" do
     it "resolves constants" do
       klass = MiniTest::Spec.spec_type(ProductDecorator)
-      klass.should == MiniTest::Rails::ActiveSupport::TestCase
+      klass.should be Draper::MiniTest::DecoratorTestCase
     end
 
     it "resolves strings" do
       klass = MiniTest::Spec.spec_type("ProductDecorator")
-      klass.should == MiniTest::Rails::ActiveSupport::TestCase
+      klass.should be Draper::MiniTest::DecoratorTestCase
     end
   end
 
   context "WidgetDecorator" do
     it "resolves constants" do
       klass = MiniTest::Spec.spec_type(WidgetDecorator)
-      klass.should == MiniTest::Rails::ActiveSupport::TestCase
+      klass.should be Draper::MiniTest::DecoratorTestCase
     end
 
     it "resolves strings" do
       klass = MiniTest::Spec.spec_type("WidgetDecorator")
-      klass.should == MiniTest::Rails::ActiveSupport::TestCase
+      klass.should be Draper::MiniTest::DecoratorTestCase
     end
   end
 
   context "decorator strings" do
     it "resolves DoesNotExistDecorator" do
       klass = MiniTest::Spec.spec_type("DoesNotExistDecorator")
-      klass.should == MiniTest::Rails::ActiveSupport::TestCase
+      klass.should be Draper::MiniTest::DecoratorTestCase
     end
 
     it "resolves DoesNotExistDecoratorTest" do
       klass = MiniTest::Spec.spec_type("DoesNotExistDecoratorTest")
-      klass.should == MiniTest::Rails::ActiveSupport::TestCase
+      klass.should be Draper::MiniTest::DecoratorTestCase
     end
 
     it "resolves Does Not Exist Decorator" do
       klass = MiniTest::Spec.spec_type("Does Not Exist Decorator")
-      klass.should == MiniTest::Rails::ActiveSupport::TestCase
+      klass.should be Draper::MiniTest::DecoratorTestCase
     end
 
     it "resolves Does Not Exist Decorator Test" do
       klass = MiniTest::Spec.spec_type("Does Not Exist Decorator Test")
-      klass.should == MiniTest::Rails::ActiveSupport::TestCase
+      klass.should be Draper::MiniTest::DecoratorTestCase
     end
   end
 
   context "non-decorators" do
     it "doesn't resolve constants" do
       klass = MiniTest::Spec.spec_type(Draper::HelperSupport)
-      klass.should == MiniTest::Spec
+      klass.should be MiniTest::Spec
     end
 
     it "doesn't resolve strings" do
       klass = MiniTest::Spec.spec_type("Nothing to see here...")
-      klass.should == MiniTest::Spec
+      klass.should be MiniTest::Spec
     end
   end
 end
